补充了爬梯子的实现逻辑, 未完成其纹理设置; 将攻击逻辑从行动模块中分离了出来

+++ core/src/zdream/rockchronicle/platform/world/LevelWorld.java
@@ -683,7 +683,7 @@ public class LevelWorld implements ITerrainStatic {
 	 * @return
 	 *   地形码
 	 */
-	public int getTerrain(int x, int y) {
+	public byte getTerrain(int x, int y) {
 		if (x < 0 || x >= currentRoom.width) {
 			return TERRAIN_SOLID;
 		}
